[BEAM-8335] Refactor IPythonLogHandler
1. Moved IPythonLogHandler from interactive_utils to interactive.utils.
2. Fixed the formatting of emit function in IPythonLogHandler.
3. Ignored index/column positioning from dataframe equivalence
assertions.
4. Removed check for jsons since it's not a dependency any more.
