+++ sdks/python/apache_beam/utils/interactive_utils.py
@@ -101,55 +101,3 @@ def _extract_pipeline_of_pvalueish(pvalueish):
   if hasattr(pvalue, 'pipeline'):
     return pvalue.pipeline
   return None
-
-
-# TODO(BEAM-8335): Move this function and the IPythonLogHandler to a util class
-# under interactive package when streaming cache changes are merged.
-def register_ipython_log_handler():
-  """Adds the IPython handler to a dummy parent logger (named
-  'apache_beam.runners.interactive') of all interactive modules' loggers so that
-  if is_in_notebook, logging displays the logs as HTML in frontends."""
-  # apache_beam.runners.interactive is not a module, thus this "root" logger is
-  # a dummy one created to hold the IPython log handler. When children loggers
-  # have propagate as True (by default) and logging level as NOTSET (by default,
-  # so the "root" logger's logging level takes effect), the IPython log handler
-  # will be triggered at the "root"'s own logging level. And if a child logger
-  # sets its logging level, it can take control back.
-  interactive_root_logger = logging.getLogger('apache_beam.runners.interactive')
-  if any([isinstance(h, IPythonLogHandler)
-          for h in interactive_root_logger.handlers]):
-    return
-  interactive_root_logger.setLevel(logging.INFO)
-  interactive_root_logger.addHandler(IPythonLogHandler())
-  # Disable the propagation so that logs emitted from interactive modules should
-  # only be handled by loggers and handlers defined within interactive packages.
-  interactive_root_logger.propagate = False
-
-
-class IPythonLogHandler(logging.Handler):
-  """A logging handler to display logs as HTML in IPython backed frontends."""
-  log_template = """
-            <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
-            <div class="alert alert-{level}">{msg}</div>"""
-
-  logging_to_alert_level_map = {
-      logging.CRITICAL: 'danger',
-      logging.ERROR: 'danger',
-      logging.WARNING: 'warning',
-      logging.INFO: 'info',
-      logging.DEBUG: 'dark',
-      logging.NOTSET: 'light'
-  }
-
-  def emit(self, record):
-    try:
-      from html import escape
-      from IPython.core.display import HTML
-      from IPython.core.display import display
-      display(
-          HTML(
-              self.log_template.format(
-                  level=self.logging_to_alert_level_map[record.levelno],
-                  msg=escape(record.msg))))
-    except ImportError:
-      pass  # NOOP when dependencies are not available.
